move beacon tests to tests folder
